requirements.txt
# This file is used by pip to install required python packages
# Usage: pip install -r requirements.txt

Flask>=2.0.1
PyJWT>=2.1.0
gunicorn==20.0.4
